requirements.txt
# Euroleague Advanced Analytics Platform — Dependencies

# Data extraction
euroleague-api>=0.0.21

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Environment variables
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
